# Conflicts:
#	common-utitlities/src/main/java/org/eea/kafka/io/KafkaSender.java

+++ dataflow-service/src/main/java/org/eea/dataflow/persistence/repository/ContributorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * The interface Contributor repository.
  */
-public interface ContributorRepository extends CrudRepository<Contributor, Integer> {
+public interface ContributorRepository extends CrudRepository<Contributor, Long> {
 
 }
